Add installation testing on Centos 10

+++ tests/vagrant/install.sh
@@ -165,33 +165,20 @@ function prepare_vm() {
   if [ -f /etc/os-release ]; then
     source /etc/os-release
     case $ID in
-      ubuntu)
-          [[ "${TEST_REPO_ENABLE}" == 'true' ]] && add-repo-deb
-          ;;
-
-      debian)
-          [ "$VERSION_CODENAME" == "bookworm" ] && apt-get update -y && apt install -y curl gnupg
-          apt-get remove postfix -y && echo "${COLOR_GREEN}☑ PREPAVE_VM: Postfix was removed${COLOR_RESET}"
-          [[ "${TEST_REPO_ENABLE}" == 'true' ]] && add-repo-deb
-          ;;
-
-      fedora)
-          [[ "${TEST_REPO_ENABLE}" == 'true' ]] && add-repo-rpm
-          if [[ "$VERSION_ID" == "40" || "$VERSION_ID" == "41" ]]; then
-              resize_fedora_disk
-          fi
-          ;;
-
-      centos)
-          [ "$VERSION_ID" == "8" ] && sed -i 's|^mirrorlist=|#&|; s|^#baseurl=http://mirror.centos.org|baseurl=http://vault.centos.org|' /etc/yum.repos.d/CentOS-*
-          [[ "${TEST_REPO_ENABLE}" == 'true' ]] && add-repo-rpm
-          yum -y install centos*-release 
-          ;;
-
+      debian | ubuntu)
+          [[ "${TEST_REPO_ENABLE}" == 'true' ]] && add-repo-deb ;;
+      centos | fedora )
+          [[ "${TEST_REPO_ENABLE}" == 'true' ]] && add-repo-rpm ;;
       *)
-          echo "${COLOR_RED}Failed to determine Linux dist${COLOR_RESET}"; exit 1
-          ;;
+          echo "${COLOR_RED}Failed to determine Linux dist${COLOR_RESET}"; exit 1 ;;
     esac
+
+    if [[ "$ID" == "debian" ]]; then
+      [ "$VERSION_CODENAME" == "bookworm" ] && apt-get update -y && apt install -y curl gnupg
+      apt-get remove postfix -y && echo "${COLOR_GREEN}☑ PREPAVE_VM: Postfix was removed${COLOR_RESET}"
+    elif [[ "$ID" == "fedora" ]]; then
+      resize_fedora_disk
+    fi
   else
       echo "${COLOR_RED}File /etc/os-release doesn't exist${COLOR_RESET}"; exit 1
   fi
